+++ src/LinkDotNet.NCronJob/NCronJobOptions.cs
@@ -12,15 +12,4 @@ public sealed class NCronJobOptions
     /// The higher the value, the longer it will take to run instant jobs.
     /// </remarks>
     public TimeSpan TimerInterval { get; set; } = TimeSpan.FromSeconds(1);
-
-    /// <summary>
-    /// Determines whether cron expressions can specify second-level precision.
-    /// </summary>
-    /// <remarks>
-    /// When enabled, cron expressions must include a seconds field, allowing for more precise scheduling.
-    /// By default, this is disabled, and cron expressions are expected to start with the minute field.
-    /// Enabling this affects scheduling granularity and may influence performance, especially for jobs
-    /// that are scheduled to run very frequently.
-    /// </remarks>
-    public bool EnableSecondPrecision { get; set; }
 }